# .md/API-CONTRACT.yaml
openapi: 3.0.3
info:
  title: Turma do Rola - Comary — Contrato de API
  version: 0.12.0
  description: >-
    Contrato publicado incrementalmente, endpoint a endpoint, assim que cada
    contrato fica estável (TASK.md Seção 1.7 / GUARDRAILS.md regra implícita
    de rastreabilidade de API) — não é preciso esperar a tarefa de Backend
    inteira terminar. Frontend/Mobile começam a consumir assim que um
    endpoint aparece aqui, mesmo que a implementação completa da tarefa
    ainda esteja em andamento. Mudar um contrato já publicado é permitido,
    mas nunca silenciosamente — toda mudança de contrato incompatível
    incrementa `info.version` e é registrada na seção "Changelog" no final
    deste arquivo.

    Duas famílias de endpoint neste sistema:
      1. **PostgREST autogerado pelo Supabase** sobre as views públicas
         curadas (`ranking_publico`, `presenca_mensal_publica`) — consumido
         **diretamente** pelo Frontend com a chave `anon` do Supabase, sem
         passar pela camada de API própria (SDD.md Seção 7.5, ADR-005). É o
         único caso em que o Frontend fala com o Supabase sem intermediação
         de um Route Handler — decisão já tomada no SDD.md, não uma exceção
         nova. Documentado aqui (paths desta seção) porque também é
         superfície pública real, exposta à internet.
      2. **Route Handlers do Next.js (App Router)**, publicados a partir de
         BE-04 em diante — toda escrita e toda leitura autenticada da área
...
servers:
...
components:
  securitySchemes:
    supabaseAnonKey:
      type: apiKey
      in: header
      name: apikey
      description: >-
        Chave pública (`anon`) do Supabase — único credential usado pelas
        páginas públicas (T02/T03). Nunca a `service role` (TASK.md Seção
        1.2/1.7, GUARDRAILS.md regra 7).
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessao_interna
      description: >-
        Cookie de sessão emitido por `POST /api/auth/login` (`httpOnly`,
        `Secure`, `SameSite=Strict`, TTL 8-12h — ADR-004, BE-04). Único
        mecanismo de autenticação da área interna (modelo binário, sem
        RBAC — SDD.md Seção 7.2): sessão válida = acesso total. Toda rota
        de escrita futura desta área (BE-06 em diante) deve declarar este
        esquema de segurança.
  responses:
    SessaoInvalida:
      description: >-
        Sem sessão válida (`middleware.ts`, BE-04, GUARDRAILS.md regra 17) —
        toda rota de escrita da área interna, exceto `/api/auth/login` e
        `/api/auth/logout`, retorna isto quando o cookie `sessao_interna`
        está ausente, expirado ou com assinatura inválida. Resposta
        reutilizável — toda rota de escrita futura (BE-06 em diante) deve
        referenciar `#/components/responses/SessaoInvalida` para o caso
        401, em vez de redefinir o formato.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroSessaoInvalida"
  schemas:
...
    LoginRequest:
      type: object
      description: Corpo de `POST /api/auth/login` (BE-04).
      properties:
        senha:
          type: string
          minLength: 1
      required: [senha]
      additionalProperties: false
    AuthErroGenerico:
      type: object
      description: >-
        Corpo de erro `401` de `POST /api/auth/login` — SEMPRE a mesma
        mensagem, esteja a falha por senha incorreta ou por bloqueio de
        rate limiting (RF-07.3, GUARDRAILS.md regra 15). Nenhum campo
        adicional (ex.: tempo restante de bloqueio) diferencia os dois
        casos — o cliente nunca consegue distinguir um do outro pela
        resposta.
      properties:
        error:
          type: string
          example: "Senha incorreta."
      required: [error]
      additionalProperties: false
    ErroRequisicaoInvalida:
      type: object
      description: >-
        Corpo de erro `400` para requisição malformada (ex.: campo `senha`
        ausente/vazio) — classe de erro diferente de "senha incorreta"/
        "bloqueado" (RF-07.3 só rege esses dois casos, não validação básica
        de input).
      properties:
        error:
          type: string
          example: "Requisição inválida."
      required: [error]
      additionalProperties: false
    ErroSessaoInvalida:
      type: object
      description: >-
        Corpo de erro `401` emitido pelo middleware de sessão
        (`middleware.ts`, BE-04) para qualquer rota de escrita sem sessão
        válida — formato distinto do erro de login (contexto diferente:
        aqui é "sessão válida ou não", não "senha certa ou errada").
      properties:
        error:
          type: string
          example: "Sessão inválida ou expirada."
      required: [error]
      additionalProperties: false
    AtletaBody:
      type: object
      description: >-
        Corpo de `POST /api/atletas` e `PUT /api/atletas/{id}` (BE-06).
        Campos em `snake_case`, mesma convenção já usada pelas respostas de
        `RankingPublicoItem`/`PresencaMensalPublicaItem` no restante deste
        contrato (decisão de consistência, TASK.md Seção 1.0 — documentada
        na nota de status de BE-06 do `TASK.md`, não escalada).
      properties:
        nome_completo:
          type: string
          minLength: 1
        apelido_exibicao:
          type: string
          minLength: 1
          description: >-
            Opcional — quando ausente/em branco, o Backend deriva o primeiro
            "token" de `nome_completo` (RF-01.2/RN-06).
        contato:
          type: string
          minLength: 1
          nullable: true
        data_nascimento:
          type: string
          format: date
          description: >-
            Obrigatório neste formulário (RF-01.1, UX-SPEC.md T04 marca com
            `*`) — não pode ser no futuro. Coluna correspondente em
            `app.atleta` é `nullable` no banco só para acomodar registro
            migrado do legado sem essa informação (RF-08.3/BE-15), não para
            tornar o campo opcional aqui.
        consentimento_responsavel_obtido:
          type: boolean
          default: false
          description: >-
            Obrigatoriamente `true` quando `data_nascimento` resultar em
            idade < 18 anos na data da requisição — caso contrário a API
            recusa com `400` (RF-01.3/RN-02).
        pontuacao_inicial:
          type: number
          minimum: 0
        confirmar_duplicidade:
          type: boolean
          default: false
          description: >-
            Quando `nome_completo` coincide (normalizado — trim/espaço/caixa)
            com o de outro atleta ativo e este campo não vier `true`, a API
            recusa com `409` (alerta de duplicidade, RF-01.5) em vez de
            gravar. O Frontend reenvia a mesma requisição com
            `confirmar_duplicidade: true` depois que o organizador confirmar
            no modal (UX-SPEC.md T04) — nunca um bloqueio definitivo, não há
            `UNIQUE` em `nome_completo` no schema (BE-02).
      required: [nome_completo, data_nascimento, pontuacao_inicial]
      additionalProperties: false
    AtletaResponse:
      type: object
      description: >-
        Resposta de `GET/POST /api/atletas` (lista/criação) e
        `GET/PUT /api/atletas/{id}` (BE-06). Nunca devolvida a cliente sem
        sessão válida (RN-01/GUARDRAILS.md regra 19, ver tag `interno`).
      properties:
        id:
          type: string
          format: uuid
        nome_completo:
          type: string
        apelido_exibicao:
          type: string
        contato:
          type: string
          nullable: true
        data_nascimento:
          type: string
          format: date
          nullable: true
        consentimento_responsavel_obtido:
          type: boolean
        pontuacao_inicial:
          type: number
        ativo:
          type: boolean
        anonimizado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Preenchido pela função `anonimizar_atleta` (ADR-011, BE-07) —
            ver `POST /api/atletas/{id}/anonimizar`.
        criado_em:
          type: string
          format: date-time
        nivel_tecnico:
          type: number
          description: >-
            RN-03 — média de pontos de eventos de jogo (gol/cartão) por
            rodada com presença (`app.atleta_nivel_tecnico`, migration
            BE-06). Fallback = `pontuacao_inicial` para atleta sem nenhuma
            presença registrada. Nunca um campo de entrada do formulário
            (RF-01.4 — sempre derivado).
        rodadas_presentes:
          type: integer
          description: Denominador usado no cálculo de `nivel_tecnico` acima.
      required:
        [
          id,
          nome_completo,
          apelido_exibicao,
          contato,
          data_nascimento,
          consentimento_responsavel_obtido,
          pontuacao_inicial,
          ativo,
          anonimizado_em,
          criado_em,
          nivel_tecnico,
          rodadas_presentes,
        ]
      additionalProperties: false
    ErroValidacaoAtleta:
      type: object
      description: >-
        Corpo de erro `400` de `POST/PUT /api/atletas*` — corpo malformado ou
        que falha `AtletaBody` (inclui o bloqueio de RF-01.3: menor de 18 sem
        `consentimento_responsavel_obtido`, reportado como um item de
        `detalhes` com `path: ["consentimento_responsavel_obtido"]`).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroDuplicidadeAtleta:
      type: object
      description: >-
        Corpo de erro `409` de `POST/PUT /api/atletas*` (RF-01.5) — emitido
        quando `nome_completo` (normalizado) coincide com o de outro atleta
        ativo e a requisição não trouxe `confirmar_duplicidade: true`.
      properties:
        error:
          type: string
          example: "duplicidade"
        atletas_duplicados:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              nome_completo:
                type: string
            required: [id, nome_completo]
      required: [error, atletas_duplicados]
      additionalProperties: false
    ErroAtletaNaoEncontrado:
      type: object
      description: Corpo de erro `404` de `GET/PUT /api/atletas/{id}` para um id inexistente.
      properties:
        error:
          type: string
          example: "Atleta não encontrado."
      required: [error]
      additionalProperties: false
    ErroAtletaJaAnonimizado:
      type: object
      description: >-
        Corpo de erro `409` de `POST /api/atletas/{id}/anonimizar` (BE-07) —
        emitido quando o atleta já foi anonimizado anteriormente
        (`anonimizado_em` já preenchido). A operação é irreversível por
        desenho (ADR-011, sem função inversa): a função PL/pgSQL
        `app.anonimizar_atleta` recusa reprocessar o mesmo atleta em vez de
        sobrescrever de novo silenciosamente.
      properties:
        error:
          type: string
          example: "Este atleta já foi anonimizado anteriormente."
      required: [error]
      additionalProperties: false
...
    LogAuditoriaItem:
      type: object
      description: >-
        Uma entrada de `app.log_auditoria` (BE-09, RF-04.4/RF-04.5) — grava
        toda correção/exclusão de rodada (BE-09) e toda anonimização de
        atleta (BE-07, ADR-011). NUNCA inclui campo de autor individual
        (RN-12/GUARDRAILS.md regra 18) — a própria tabela não tem essa
        coluna. `valores_antes`/`valores_depois` são objetos livres (jsonb)
        cujo formato depende de `tipo_evento`: para `"correcao"`, contêm
        `status`/`eventos`/`pontos_acumulados` (e `ajuste_aplicado` só em
        `valores_depois`) da participação corrigida; para `"estorno"`,
        contêm `status` da rodada e, em `valores_depois`,
        `atletas_afetados`/`pontos_revertidos`/`substituicoes_vinculadas`;
        para `"anonimizacao"` (BE-07), `valores_antes` contém SOMENTE
        marcadores redigidos (`"[REDACTED]"`), nunca o dado pessoal real.
      properties:
        id:
          type: string
          format: uuid
        rodada_id:
          type: string
          format: uuid
          nullable: true
          description: Preenchido para `tipo_evento` `correcao`/`estorno`; nulo para `anonimizacao`.
        atleta_id:
          type: string
          format: uuid
          nullable: true
          description: Preenchido para `tipo_evento` `correcao`/`anonimizacao`; nulo para `estorno` (afeta a rodada inteira).
        tipo_evento:
          type: string
          enum: [correcao, estorno, anonimizacao]
        ocorrido_em:
          type: string
          format: date-time
        valores_antes: {}
        valores_depois: {}
      required:
        [id, rodada_id, atleta_id, tipo_evento, ocorrido_em, valores_antes, valores_depois]
      additionalProperties: false
    ErroConsultaLogAuditoria:
      type: object
      description: >-
        Corpo de erro `400` de `GET /api/log-auditoria` — `limit` inválido
        (não numérico, ≤ 0, ou acima do teto máximo de 200).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
...
paths:
...
  /api/auth/login:
    post:
      tags: [interno]
      summary: Login com a senha única compartilhada (T01, ADR-004, BE-04)
      description: >-
        Único endpoint de escrita isento do middleware de sessão — é por
        aqui que uma sessão é obtida pela primeira vez. Rate limiting de 5
        tentativas erradas por IP em 15 minutos, com backoff exponencial
        (tabela `app.tentativa_login`, RNF-03/RNF-04) — a resposta de erro é
        idêntica (mesmo status HTTP, mesmo corpo) esteja a falha por senha
        incorreta ou por bloqueio de rate limiting (RF-07.3, GUARDRAILS.md
        regra 15).
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: >-
            Senha correta — sessão emitida via cookie `sessao_interna`
            (`httpOnly`, `SameSite=Strict`, `Max-Age` entre 8h e 12h;
            `Secure` reflete o protocolo real da requisição — sempre `true`
            em produção via Vercel/HTTPS, `false` apenas em `next dev` local
            sobre HTTP, decisão de detalhe documentada em
            `src/modules/autenticacao/session-cookie.ts`).
          headers:
            Set-Cookie:
              schema:
                type: string
                example: >-
                  sessao_interna=<token>; Max-Age=36000; Path=/; HttpOnly;
                  SameSite=Strict; Secure
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]
                additionalProperties: false
        "400":
          description: >-
            Corpo malformado (ex.: campo `senha` ausente/vazio) — classe de
            erro diferente de "senha incorreta"/"bloqueado".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroRequisicaoInvalida"
        "401":
          description: >-
            Senha incorreta OU bloqueado por rate limiting — resposta
            idêntica nos dois casos (RF-07.3).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErroGenerico"
...
  /api/atletas:
    get:
      tags: [interno]
      summary: Lista atletas (BE-06, RF-01)
      description: >-
        Exige sessão válida mesmo sendo leitura (ver tag `interno`) —
        devolve `contato`/`data_nascimento`. Cada item inclui o nível
        técnico (RN-03) já calculado (`nivel_tecnico`/`rodadas_presentes`).
        Sem paginação nesta versão (porte do grupo — dezenas de atletas, não
        milhares, TASK.md Seção 1.0).
      operationId: getAtletas
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de atletas, ordenada por `nome_completo` asc.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtletaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Cadastra um atleta (BE-06, RF-01.1)
      description: >-
        Inicializa o saldo acumulado do atleta com `pontuacao_inicial`
        (RF-01.1/RN-10). Bloqueia com `400` se `data_nascimento` resultar em
        idade < 18 anos e `consentimento_responsavel_obtido` não vier `true`
        (RF-01.3/RN-02). Alerta de duplicidade de `nome_completo` (RF-01.5)
        — ver `409` abaixo.
      operationId: postAtletas
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaBody"
      responses:
        "201":
          description: Atleta criado. `nivel_tecnico` = `pontuacao_inicial` (nenhuma presença ainda, fallback RN-03).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "400":
          description: >-
            Corpo malformado/inválido — inclui o bloqueio de RF-01.3 (menor
            de 18 sem consentimento marcado) como um item de `detalhes` com
            `path: ["consentimento_responsavel_obtido"]`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoAtleta"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "409":
          description: >-
            `nome_completo` (normalizado) coincide com o de outro atleta
            ativo e a requisição não trouxe `confirmar_duplicidade: true`
            (RF-01.5) — nenhum registro foi criado. Reenviar a mesma
            requisição com `confirmar_duplicidade: true` para confirmar e
            criar mesmo assim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeAtleta"
  /api/atletas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [interno]
      summary: Busca um atleta por id (BE-06)
      description: >-
        Exige sessão válida mesmo sendo leitura (ver tag `interno`) — mesmo
        motivo de `GET /api/atletas`.
      operationId: getAtletaPorId
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Atleta encontrado, com nível técnico (RN-03) calculado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: Nenhum atleta com este id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaNaoEncontrado"
    put:
      tags: [interno]
      summary: Atualiza um atleta já cadastrado (BE-06, RF-01.6)
      description: >-
        Substitui os campos editáveis (mesmo corpo de `POST /api/atletas`) —
        nunca altera `ativo`/`anonimizado_em` (mecanismo próprio de BE-07,
        `anonimizar_atleta`, ADR-011, fora do escopo desta rota). Reaplica a
        mesma checagem de idade/consentimento (RF-01.3) e de duplicidade de
        nome (RF-01.5) usada na criação, excluindo o próprio id da checagem
        de duplicidade.
      operationId: putAtletaPorId
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaBody"
      responses:
        "200":
          description: Atleta atualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoAtleta"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: Nenhum atleta com este id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaNaoEncontrado"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeAtleta"
  /api/atletas/{id}/anonimizar:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [interno]
      summary: Anonimiza dado pessoal do atleta a pedido do titular (BE-07, ADR-011, LGPD Art. 18)
      description: >-
        Aciona a função PL/pgSQL `app.anonimizar_atleta`, que roda dentro de
        uma única transação: sobrescreve `nome_completo` (sempre
        `"Atleta anonimizado"`), `apelido_exibicao` (sempre
        `"Atleta #" + primeiros 8 caracteres do id`), `contato` e
        `data_nascimento` (sempre `null`), marca `ativo=false` e
        `anonimizado_em`, desativa toda `restricao_obrigatoria` (RF-05.5)
        onde o atleta seja `atleta_a_id`/`atleta_b_id`, e grava uma entrada
        em `log_auditoria` (`tipo_evento="anonimizacao"`) com `valores_antes`
        contendo **apenas marcadores redigidos** (`"[REDACTED]"`), nunca o
        dado pessoal real (GUARDRAILS.md regra 20). Nunca altera
        `lancamento_pontos`/`participacao_rodada`/`time_atleta`/
        `substituicao` — o saldo/histórico agregado do atleta permanece
        íntegro e consultável pelo mesmo id. Ação **irreversível por
        desenho** (ADR-011, sem função inversa) — nenhum corpo de
        requisição é exigido; a confirmação de dupla etapa
        (`TypedConfirmationModal`, "digite ANONIMIZAR") é responsabilidade
        de fluxo do Frontend (UX-SPEC.md T04), não um contrato desta API.
      operationId: postAnonimizarAtleta
      security:
        - sessionCookie: []
      responses:
        "200":
          description: >-
            Atleta anonimizado. Resposta reflete os campos já sobrescritos
            (`contato`/`data_nascimento` nulos, `ativo=false`,
            `anonimizado_em` preenchido) — `nivel_tecnico`/
            `rodadas_presentes` inalterados (histórico preservado).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: Nenhum atleta com este id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaNaoEncontrado"
        "409":
          description: Atleta já anonimizado anteriormente (irreversível, ADR-011).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaJaAnonimizado"
...
  /api/log-auditoria:
    get:
      tags: [interno]
      summary: Consulta o log de auditoria de correções/exclusões e anonimizações (BE-09, RF-04.5)
      description: >-
        Lista `app.log_auditoria` ordenado do mais recente ao mais antigo
        (`ocorrido_em desc`, RF-04.5) — grava toda correção/exclusão de
        rodada (BE-09) e toda anonimização de atleta (BE-07, ADR-011).
        NUNCA inclui campo de autor individual (RN-12/GUARDRAILS.md regra
        18) — a própria tabela não tem essa coluna. Leitura interna
        protegida por sessão mesmo sendo `GET` (RF-07.1 lista "histórico/
        correção" entre as ações que exigem a senha interna) —
        `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`, mesmo padrão
        já usado por `GET /api/atletas` (BE-06).
      operationId: getLogAuditoria
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: >-
            Teto de linhas retornadas (decisão de detalhe, TASK.md Seção
            1.0 — não exigido literalmente pelo critério de aceite; evita
            crescimento não controlado da resposta, já que
            `app.log_auditoria` é retido indefinidamente sem expurgo
            automático, RNF-06).
      responses:
        "200":
          description: Lista ordenada do mais recente ao mais antigo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogAuditoriaItem"
        "400":
          description: "`limit` inválido (não numérico, ≤ 0, ou acima de 200)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroConsultaLogAuditoria"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
...